feat: добавлена реализация методов для фильмов

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -20,6 +20,6 @@ public class User {
     private String login;
     private String name;
     private LocalDate birthday;
-    private final Set<Long> friends = new HashSet<>();
+    private final Set<Integer> friends = new HashSet<>();
 
 }
